# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Product:
    properties:
      brand_id:
        type: integer
      category_id:
        type: integer
      code:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      order_number:
        type: integer
      price:
        type: number
      promotion_id:
        type: integer
      status:
        type: integer
      supplier_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.SingIn:
    properties:
      password:
        type: string
      phone:
        type: string
    type: object
...
  utils.DataResponse:
    properties:
      error_code:
        type: integer
      message: {}
    type: object
host: localhost:5005
info:
  contact: {}
  description: Api INSIGHT project
  title: Insight API SERVICE
  version: 1.0.0
paths:
  /auth/change-password:
    put:
      consumes:
      - application/json
      description: changePassword
      operationId: changePassword
      parameters:
      - description: Заполните поля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: changePassword
      tags:
      - Auth
  /auth/log-out:
    put:
      consumes:
      - application/json
      description: Роут для выхода из системы
      operationId: logoutHandler
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: logoutHandler
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: sign-in in account
      operationId: loginHandler
      parameters:
      - description: user info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SingIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      summary: loginHandler
      tags:
      - Auth
  /auth/refresh-token:
    post:
      description: Роут для обвноляения токена
      operationId: refreshToken
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: refreshToken
      tags:
      - Auth
  /products/by-id:
    get:
      consumes:
      - application/json
      description: Просмотр магазина
      operationId: getProduct
      parameters:
      - description: Введите данные
        in: query
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getProduct
      tags:
      - Products
  /products/edit:
    put:
      consumes:
      - application/json
      description: Редактирование продукта
      operationId: editProduct
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: editProduct
      tags:
      - Products
  /products/list:
    get:
      consumes:
      - application/json
      description: Просмотр списока всех продуктов
      operationId: getAllProducts
      parameters:
      - description: Введите данные
        in: query
        name: page
        required: true
        type: string
      - description: Введите данные
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getAllProducts
      tags:
      - Products
  /products/new:
    post:
      consumes:
      - application/json
      description: Добавление нового продукта
      operationId: addNewProduct
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: addNewProduct
      tags:
      - Products
  /products/rm:
    delete:
      consumes:
      - application/json
      description: Удаление продукта
      operationId: deleteProduct
      parameters:
      - description: Введите данные
        in: query
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: deleteProduct
      tags:
      - Products
...
  /users/by-id:
    get:
      consumes:
      - application/json
      description: Просмотр пользователя
      operationId: getUserById
      parameters:
      - description: Введите данные
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getUserById
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"